backend/requirements.txt
# Core Framework
fastapi
uvicorn
pydantic
pydantic-settings
websockets

# Database
sqlalchemy
alembic
psycopg2-binary

# LangChain & LangGraph
langgraph
langchain
langchain-core
langchain-openai
langchain-community
langsmith

# OpenAI (for audio & vision)
openai

# Multimodal Support
pillow

# Environment & Configuration
python-dotenv
python-multipart

# Utilities
requests

# Voice Chat (Optional - only needed for voice features)
livekit
